feat: reorganize files and folders
