+++ app/not-found.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import NotFound from "@/appComponents/NotFound";
+import NotFound from "@/appComponents/structure/NotFound";
 
 export default function Custom404() {
   return <NotFound />;
