+++ app/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import HomeLayout from "@/appComponents/HomeLayout";
+import HomeLayout from "@/appComponents/layouts/HomeLayout";
 import { ARDI } from "@/lib/constants";
 
 export async function generateMetadata() {
