--- app/ui/Stats.module.css
@@ -1,206 +1,0 @@
-.stats {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-}
-
-.stats a:hover {
-  cursor: pointer;
-}
-
-.stats > a > img {
-  width: auto;
-  height: 70px;
-  border-radius: var(--border-radius);
-  z-index: 10;
-  position: relative;
-  transition: all 0.3s;
-}
-
-.stats > a > img:hover {
-  transform: rotate(-2deg);
-}
-
-.stats > a:before {
-  background: rgba(var(--color-space-darker));
-  border-radius: var(--border-radius);
-  content: "";
-  height: 70px;
-  margin-left: -10px;
-  margin-top: 10px;
-  position: absolute;
-  transform: rotate(85deg);
-  transition: all 0.3s;
-  width: 70px;
-}
-
-.stats > a:hover:before {
-  margin-left: -5px;
-  margin-top: 5px;
-  transform: rotate(90deg);
-}
-
-.stats > .indicators {
-  display: grid;
-  grid-template-rows: 20px 25px 20px;
-}
-
-.stats > .indicators > .hearts {
-  display: flex;
-  gap: 10px;
-  height: 20px;
-}
-
-.stats > .indicators > .info {
-  display: flex;
-  flex-wrap: nowrap;
-  gap: 10px;
-  cursor: help;
-}
-
-.stats > .indicators > .info p {
-  margin: 0;
-}
-
-.stats .honeyYellow {
-  color: rgba(var(--color-honey-yellow));
-}
-
-.indicators > .experience > progress {
-  width: 100%;
-  height: 15px;
-  border: 2px solid white;
-  border-radius: calc(var(--border-radius) * 5);
-}
-
-.indicators > .regular > progress::-webkit-progress-bar {
-  background-color: rgba(255, 220, 36, 0.2);
-  border-radius: var(--border-radius);
-}
-
-.indicators > .birthday > progress {
-  animation: glow 1s infinite;
-  animation-timing-function: linear;
-}
-
-.indicators > .regular > progress::-webkit-progress-value {
-  background-color: rgba(var(--color-honey-yellow));
-  border-top-left-radius: var(--border-radius);
-  border-bottom-left-radius: var(--border-radius);
-}
-
-.indicators > .birthday > progress::-webkit-progress-bar {
-  border-radius: var(--border-radius);
-}
-
-.indicators > .birthday > progress::-webkit-progress-value {
-  border-radius: var(--border-radius);
-  background: linear-gradient(
-    to right,
-    rgba(var(--color-honey-yellow)) 0%,
-    rgba(var(--color-red-rum)) 50%,
-    rgba(var(--color-kokiri-green)) 75%,
-    rgba(var(--color-calm-blue)) 100%
-  );
-}
-
-@media (max-width: 768px) {
-  .stats {
-    height: 70px;
-    z-index: 20;
-  }
-
-  .stats > a {
-    display: flex;
-  }
-
-  .stats > a > img {
-    height: 40px;
-  }
-
-  .stats > a > img:hover {
-    margin-left: 0;
-    margin-top: -5px;
-    transform: none;
-  }
-
-  .stats > a:before {
-    background: rgba(var(--color-space-darker));
-    border-radius: var(--border-radius);
-    content: "";
-    height: 70px;
-    margin-left: -10px;
-    margin-top: 10px;
-    position: absolute;
-    transform: none;
-    transition: all 0.3s;
-    width: 70px;
-    rotate: none;
-  }
-
-  .stats > a:before {
-    background: black;
-    border-radius: 3px;
-    content: "";
-    height: 40px;
-    margin-left: 3px;
-    margin-top: 3px;
-    position: absolute;
-    transform: none;
-    transition: all 0.3s;
-    width: 35px;
-  }
-
-  .stats > a:hover:before {
-    margin-left: 3px;
-    margin-top: 1px;
-    transform: rotate(0);
-  }
-
-  .stats > .indicators {
-    display: grid;
-    grid-template-areas:
-      "hearts experience"
-      "info info";
-    grid-template-rows: initial;
-    grid-template-columns: 1fr 1fr;
-    align-items: center;
-    grid-column-gap: 10px;
-  }
-
-  .stats > .indicators > .hearts {
-    grid-area: hearts;
-    display: flex;
-    gap: 5px;
-    height: 20px;
-  }
-  .indicators > .experience > progress {
-    grid-area: experience;
-    width: 100%;
-    height: 15px;
-    border: 2px solid white;
-    border-radius: calc(var(--border-radius) * 5);
-  }
-
-  .stats > .indicators > .info {
-    grid-area: info;
-  }
-}
-
-@keyframes glow {
-  0% {
-    border: 2px solid white;
-    background-position-x: 0;
-  }
-
-  50% {
-    border: 2px solid #ffd812;
-    box-shadow: #fff41a 0 0 10px;
-    background-position-x: 400px;
-  }
-
-  100% {
-    border: 2px solid white;
-    background-position-x: 0;
-  }
-}
