+++ app/ui/layouts/HomeLayout.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
-import BasicLayout from "@/appComponents/BasicLayout";
-import Posts from "@/appComponents/Posts";
-import Welcome from "@/appComponents/Welcome";
+import BasicLayout from "@/appComponents/layouts/BasicLayout";
+import Posts from "@/appComponents/general/Posts";
+import Welcome from "@/appComponents/general/Welcome";
 import { getSortedPostsData } from "@/lib/getPostData";
 
-import global from "../ui/Global.module.css";
+import global from "../../ui/general/Global.module.css";
 
 const Home = () => {
   const allPostsData = getSortedPostsData(5);
