+++ app/ui/structure/MainNavigation.tsx
@@ -2,9 +2,9 @@
 
 import React from "react";
 
-import Stats from "@/appComponents/Stats";
+import Stats from "@/appComponents/general/Stats";
 
-import LinkItems from "./LinkItems";
+import LinkItems from "../../ui/general/LinkItems";
 import styles from "./MainNavigation.module.css";
 
 const MainNavigation = () => {
